+++ src/main/java/io/github/a5b84/helditeminfo/mixin/HeldItemTooltipMixin.java
@@ -21,6 +21,9 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 import java.util.Collections;
 import java.util.List;
 
+import static io.github.a5b84.helditeminfo.Mod.FONT_HEIGHT;
+import static io.github.a5b84.helditeminfo.Mod.config;
+
 @Mixin(InGameHud.class)
 public abstract class HeldItemTooltipMixin extends DrawableHelper {
 
@@ -31,10 +34,6 @@ public abstract class HeldItemTooltipMixin extends DrawableHelper {
     @Shadow private int scaledHeight;
 
 
-
-    @Unique private static final int FONT_HEIGHT = 9;
-    @Unique private static final int LINE_HEIGHT = FONT_HEIGHT - 1;
-    @Unique private static final float OFFSET_PER_EXTRA_LINE = LINE_HEIGHT - .334f;
 
     @Unique private List<InfoLine> info = Collections.emptyList();
     @Unique private int y;
@@ -53,7 +52,7 @@ public abstract class HeldItemTooltipMixin extends DrawableHelper {
     )
     public void onBeforeRenderHeldItemTooltip(CallbackInfo ci) {
         // 50 = 32 (hotbar) + 14 (health & xp) + 4 (spacing) (?)
-        y = scaledHeight - 50 - FONT_HEIGHT - (int) (OFFSET_PER_EXTRA_LINE * (info.size() - 1));
+        y = scaledHeight - 50 - FONT_HEIGHT - (int) ((config.lineHeight() - config.offsetPerExtraLine()) * (info.size() - 1));
         //noinspection ConstantConditions
         if (!this.client.interactionManager.hasStatusBars()) y += 14;
     }
@@ -80,7 +79,7 @@ public abstract class HeldItemTooltipMixin extends DrawableHelper {
         fill(
             stack,
             (scaledWidth - maxWidth) / 2 - 2, y - 2,
-            (scaledWidth + maxWidth) / 2 + 2, y + (LINE_HEIGHT * info.size()) + 2,
+            (scaledWidth + maxWidth) / 2 + 2, y + (config.lineHeight() * info.size()) + 2,
             color
         );
     }
@@ -93,10 +92,11 @@ public abstract class HeldItemTooltipMixin extends DrawableHelper {
         at = @At(value = "INVOKE", target = "net/minecraft/client/font/TextRenderer.drawWithShadow(Lnet/minecraft/client/util/math/MatrixStack;Lnet/minecraft/text/Text;FFI)I")
     )
     private int drawTextProxy(TextRenderer fontRenderer, MatrixStack stack, Text name, float _x, float _y, int color) {
+        final int lineHeight = config.lineHeight();
         for (final InfoLine line : info) {
             final int x = (scaledWidth - line.width) / 2;
             fontRenderer.drawWithShadow(stack, line.text, x, y, color);
-            y += LINE_HEIGHT;
+            y += lineHeight;
         }
 
         return 0;
@@ -106,17 +106,18 @@ public abstract class HeldItemTooltipMixin extends DrawableHelper {
 
     @Inject(method = "tick", at = @At("HEAD"))
     public void onBeforeTick(CallbackInfo ci) {
-        if (this.client.player != null) {
-            stackBeforeTick = currentStack;
-        }
+        stackBeforeTick = currentStack;
     }
 
     /** Rebuilds the tooltip */
     @Inject(method = "tick", at = @At("RETURN"))
     public void onAfterTick(CallbackInfo ci) {
-        if (this.client.player == null
-                || currentStack == stackBeforeTick
-                || currentStack.isEmpty()) {
+        if (client.player == null || currentStack == stackBeforeTick) {
+            return;
+        }
+
+        if (currentStack.isEmpty()) {
+            info = Collections.emptyList();
             return;
         }
 
@@ -128,7 +129,7 @@ public abstract class HeldItemTooltipMixin extends DrawableHelper {
 
         info = InfoBuilder.toInfoLines(newInfo);
         maxWidth = -1;
-        heldItemTooltipFade = 40 + 4 * (info.size() - 1);
+        heldItemTooltipFade = (int) (20 * (config.baseFadeDuration() + config.fadeDurationPerExtraLine() * (info.size() - 1)));
     }
 
 }