+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ### Unreleased
+- Added a config screen
+    - The config screen is only available if both [Mod Menu](https://www.curseforge.com/minecraft/mc-mods/modmenu) and [Cloth Config](https://www.curseforge.com/minecraft/mc-mods/cloth-config) (4.9 or higher) are installed
 - Now requires Fabric API
 
 ### [1.1.3]